Step to make Lidar classes compatible with OF-6
The lidar classes scanningLidar and spinnerLidar are setup to be compatible with
OpenFOAM-6, but one issue with compatibility remains (see below)
The changes in this commit include:
 - class inherits from functionObject class. This means that the <name>FunctionObject
    files are no longer needed
 - name_ is inherited from parent class, so declaration is removed and
    name() function is used where needed
 - some standard functions like write(), read(), execute(), ... have bool as return type

There remains one issue with compatibility, i.e.,
the function Foam::Random::vector01() no longer exists
