+++ src/functionObjects/utilities/scanningLidar/scanningLidar.H
@@ -22,25 +22,27 @@ License
     along with OpenFOAM.  If not, see <http://www.gnu.org/licenses/>.
 
 Class
-    Foam::spinnerLidar
+    Foam::scanningLidar
 
 Group
     grpUtilitiesFunctionObjects
 
 Description
-    This function models the scan pattern of DTU's spinner lidar.  It
-    samples the flow field and writes the information to file at a pre-
-    specified frequency.
+    This function models the scan pattern for a generic scanning lidar.
+    The scan pattern and scan times are user inputs.  The code repeats the
+    scans and samples the velocity field and writes the data to file as
+    the solver runs.
 
 SourceFiles
-    spinnerLidar.C
-    IOspinnerLidar.H
+    scanningLidar.C
+    IOscanningLidar.H
 
 \*---------------------------------------------------------------------------*/
 
-#ifndef spinnerLidar_H
-#define spinnerLidar_H
+#ifndef scanningLidar_H
+#define scanningLidar_H
 
+#include "functionObject.H"
 #include "fvCFD.H"
 #include "Random.H"
 
@@ -56,15 +58,14 @@ class dictionary;
 class mapPolyMesh;
 
 /*---------------------------------------------------------------------------*\
-                       Class spinnerLidar Declaration
+                       Class scanningLidar Declaration
 \*---------------------------------------------------------------------------*/
 
-class spinnerLidar
+class scanningLidar
+:
+    public functionObject
 {
     // Private data
-
-        //- Name of this set of spinnerLidar objects
-        word name_;
 
         //- Reference to the mesh database
         const fvMesh& mesh_;
@@ -101,9 +102,6 @@ class spinnerLidar
         //- Output file pointer.
         autoPtr<OFstream> outputFile;
 
-        //- Local domain bounding box.
-        const pointField& meshPoints;
-
         //- A list that defines the scan pattern.  The first column is time
         //  from zero to the completion of the pattern.  The remaining columns
         //  give the vector that the beam points along at that given time.
@@ -111,38 +109,8 @@ class spinnerLidar
         List<vector> beamScanPatternVector;
         label beamScanPatternI;
 
-        //- Does the scan pattern repeat.  Depending on the motor speed, sampling
-        //  rates, etc., the scan segments may or may not repeat.  If they do not
-        //  repeat, then for each time through the scan pattern, the interpolation
-        //  cell locations need to be relocated, which is time consuming.  If
-        //  scanRepeats is set to true, the scan segments are forced to repeat 
-        //  location (even if that means the beam has to backtrack slightly to
-        //  begin a new scan), and the relocating of interpolation cells is not
-        //  required.
-        bool scanRepeats;
-
         //- Time between scans.
         scalar timeBetweenScans;
-
-        //- DTU Spinner Lidar specific input parameters
-        scalar motorRPM;
-        scalar oneSecScanMotorRPM;
-        scalar gearRatioMotorToPrism1;
-        scalar gearRatioPrism1ToPrism2;
-        scalar prism1Angle;
-        scalar prism2Angle;
-        scalar sampleRate;
-        scalar samplePeriod;
-        label samplesPerSpectrum;
-        label spectraPerAverageSample;
-        scalar averageSampleRate;
-        scalar averageSamplePeriod;
-        label virtualSamplesPerAverageSample;
-        scalar virtualSamplePeriod;
-        scalar scanPeriod;
-        label nBeamPoints;
-        label nAverageSamples;
-        label nVirtualSamples;
 
         //- Origin of lidar beam.
         vector beamOrigin;
@@ -162,29 +130,16 @@ class spinnerLidar
         //  Use the right-hand rule.
         vector beamRotationAxis;
         vector beamElevationAxis;
-        vector beamElevationAxisOriginal;
 
         //- Value to perturb lidar sampling locations by to break ties when 
         //  searching for control processor.
         scalar perturb;
-
-        //- Last position of the prisms and beam.  At start time, these are pointed horizontally
-        //  to the right as viewed from the lidar looking downstream.  For prism 1, this is
-        //  downsteam; for prism 2, this is the prism 1 angle; for the beam, this is the angle of the
-        //  the combined prisms.
-        vector lastPrism1Axis;
-        vector lastPrism2Axis;
-        vector lastBeamAxis;
 
         //- Sampling points.
         List<List<vector> > samplePoints;
 
         //- Sampled winds.
         List<vector> sampledWindVectors;
-
-        //- Prism rotation rates.
-        scalar rr1;
-        scalar rr2;
 
         //- Perturbation vector for breaking processor ties.
         List<List<vector> > perturbVectors;
@@ -207,23 +162,23 @@ class spinnerLidar
     // Private Member Functions
 
         //- Disallow default bitwise copy construct
-        spinnerLidar(const spinnerLidar&);
+        scanningLidar(const scanningLidar&);
 
         //- Disallow default bitwise assignment
-        void operator=(const spinnerLidar&);
+        void operator=(const scanningLidar&);
 
 
 public:
 
     //- Runtime type information
-    TypeName("spinnerLidar");
+    TypeName("scanningLidar");
 
 
     // Constructors
 
         //- Construct for given objectRegistry and dictionary.
         //  Allow the possibility to load fields from files
-        spinnerLidar
+        scanningLidar
         (
             const word& name,
             const objectRegistry&,
@@ -233,25 +188,16 @@ public:
 
 
     //- Destructor
-    virtual ~spinnerLidar();
+    virtual ~scanningLidar();
 
 
     // Member Functions
 
-        //- Return name of the set of spinnerLidar
-        virtual const word& name() const
-        {
-            return name_;
-        }
-
-        //- Read the spinnerLidar data
-        virtual void read(const dictionary&);
-
-        //- Reset the beam position
-        virtual void resetAxes();
+        //- Read the scanningLidar data
+        virtual bool read(const dictionary&);
 
         //- Create the lidar sampling beams
-        virtual void createScanPattern();
+        virtual void createBeams();
 
         //- Find the processor and cell that contains the sampling point.
         virtual void findControlProcAndCell();
@@ -266,30 +212,19 @@ public:
         virtual void rotateLidar();
 
         //- Execute, currently does nothing
-        virtual void execute();
+        virtual bool execute();
 
         //- Execute at the final time-loop, currently does nothing
-        virtual void end();
-
-        //- Called when time was set at the end of the Time::operator++
-        virtual void timeSet();
+        virtual bool end();
 
         //- Standard write function.
-        virtual void write();
+        virtual bool write();
 
         //- Custom write function only called after a full scan.
-        virtual void writeBeamDataFormatted();
+        virtual void writeBeamData();
 
         //- Write out variables for debugging
         virtual void writeVariables();
-
-        //- Update for changes of mesh
-        virtual void updateMesh(const mapPolyMesh&)
-        {}
-
-        //- Update for changes of mesh
-        virtual void movePoints(const polyMesh&)
-        {}
 };
