+++ src/functionObjects/utilities/spinnerLidar/spinnerLidar.H
@@ -22,26 +22,26 @@ License
     along with OpenFOAM.  If not, see <http://www.gnu.org/licenses/>.
 
 Class
-    Foam::scanningLidar
+    Foam::spinnerLidar
 
 Group
     grpUtilitiesFunctionObjects
 
 Description
-    This function models the scan pattern for a generic scanning lidar.
-    The scan pattern and scan times are user inputs.  The code repeats the
-    scans and samples the velocity field and writes the data to file as
-    the solver runs.
+    This function models the scan pattern of DTU's spinner lidar.  It
+    samples the flow field and writes the information to file at a pre-
+    specified frequency.
 
 SourceFiles
-    scanningLidar.C
-    IOscanningLidar.H
+    spinnerLidar.C
+    IOspinnerLidar.H
 
 \*---------------------------------------------------------------------------*/
 
-#ifndef scanningLidar_H
-#define scanningLidar_H
+#ifndef spinnerLidar_H
+#define spinnerLidar_H
 
+#include "functionObject.H"
 #include "fvCFD.H"
 #include "Random.H"
 
@@ -57,15 +57,14 @@ class dictionary;
 class mapPolyMesh;
 
 /*---------------------------------------------------------------------------*\
-                       Class scanningLidar Declaration
+                       Class spinnerLidar Declaration
 \*---------------------------------------------------------------------------*/
 
-class scanningLidar
+class spinnerLidar
+:
+    public functionObject
 {
     // Private data
-
-        //- Name of this set of scanningLidar objects
-        word name_;
 
         //- Reference to the mesh database
         const fvMesh& mesh_;
@@ -102,6 +101,9 @@ class scanningLidar
         //- Output file pointer.
         autoPtr<OFstream> outputFile;
 
+        //- Local domain bounding box.
+        const pointField& meshPoints;
+
         //- A list that defines the scan pattern.  The first column is time
         //  from zero to the completion of the pattern.  The remaining columns
         //  give the vector that the beam points along at that given time.
@@ -109,8 +111,38 @@ class scanningLidar
         List<vector> beamScanPatternVector;
         label beamScanPatternI;
 
+        //- Does the scan pattern repeat.  Depending on the motor speed, sampling
+        //  rates, etc., the scan segments may or may not repeat.  If they do not
+        //  repeat, then for each time through the scan pattern, the interpolation
+        //  cell locations need to be relocated, which is time consuming.  If
+        //  scanRepeats is set to true, the scan segments are forced to repeat 
+        //  location (even if that means the beam has to backtrack slightly to
+        //  begin a new scan), and the relocating of interpolation cells is not
+        //  required.
+        bool scanRepeats;
+
         //- Time between scans.
         scalar timeBetweenScans;
+
+        //- DTU Spinner Lidar specific input parameters
+        scalar motorRPM;
+        scalar oneSecScanMotorRPM;
+        scalar gearRatioMotorToPrism1;
+        scalar gearRatioPrism1ToPrism2;
+        scalar prism1Angle;
+        scalar prism2Angle;
+        scalar sampleRate;
+        scalar samplePeriod;
+        label samplesPerSpectrum;
+        label spectraPerAverageSample;
+        scalar averageSampleRate;
+        scalar averageSamplePeriod;
+        label virtualSamplesPerAverageSample;
+        scalar virtualSamplePeriod;
+        scalar scanPeriod;
+        label nBeamPoints;
+        label nAverageSamples;
+        label nVirtualSamples;
 
         //- Origin of lidar beam.
         vector beamOrigin;
@@ -130,16 +162,29 @@ class scanningLidar
         //  Use the right-hand rule.
         vector beamRotationAxis;
         vector beamElevationAxis;
+        vector beamElevationAxisOriginal;
 
         //- Value to perturb lidar sampling locations by to break ties when 
         //  searching for control processor.
         scalar perturb;
+
+        //- Last position of the prisms and beam.  At start time, these are pointed horizontally
+        //  to the right as viewed from the lidar looking downstream.  For prism 1, this is
+        //  downsteam; for prism 2, this is the prism 1 angle; for the beam, this is the angle of the
+        //  the combined prisms.
+        vector lastPrism1Axis;
+        vector lastPrism2Axis;
+        vector lastBeamAxis;
 
         //- Sampling points.
         List<List<vector> > samplePoints;
 
         //- Sampled winds.
         List<vector> sampledWindVectors;
+
+        //- Prism rotation rates.
+        scalar rr1;
+        scalar rr2;
 
         //- Perturbation vector for breaking processor ties.
         List<List<vector> > perturbVectors;
@@ -162,23 +207,23 @@ class scanningLidar
     // Private Member Functions
 
         //- Disallow default bitwise copy construct
-        scanningLidar(const scanningLidar&);
+        spinnerLidar(const spinnerLidar&);
 
         //- Disallow default bitwise assignment
-        void operator=(const scanningLidar&);
+        void operator=(const spinnerLidar&);
 
 
 public:
 
     //- Runtime type information
-    TypeName("scanningLidar");
+    TypeName("spinnerLidar");
 
 
     // Constructors
 
         //- Construct for given objectRegistry and dictionary.
         //  Allow the possibility to load fields from files
-        scanningLidar
+        spinnerLidar
         (
             const word& name,
             const objectRegistry&,
@@ -188,22 +233,19 @@ public:
 
 
     //- Destructor
-    virtual ~scanningLidar();
+    virtual ~spinnerLidar();
 
 
     // Member Functions
 
-        //- Return name of the set of scanningLidar
-        virtual const word& name() const
-        {
-            return name_;
-        }
+        //- Read the spinnerLidar data
+        virtual bool read(const dictionary&);
 
-        //- Read the scanningLidar data
-        virtual void read(const dictionary&);
+        //- Reset the beam position
+        virtual void resetAxes();
 
         //- Create the lidar sampling beams
-        virtual void createBeams();
+        virtual void createScanPattern();
 
         //- Find the processor and cell that contains the sampling point.
         virtual void findControlProcAndCell();
@@ -218,30 +260,19 @@ public:
         virtual void rotateLidar();
 
         //- Execute, currently does nothing
-        virtual void execute();
+        virtual bool execute();
 
         //- Execute at the final time-loop, currently does nothing
-        virtual void end();
-
-        //- Called when time was set at the end of the Time::operator++
-        virtual void timeSet();
+        virtual bool end();
 
         //- Standard write function.
-        virtual void write();
+        virtual bool write();
 
         //- Custom write function only called after a full scan.
-        virtual void writeBeamData();
+        virtual void writeBeamDataFormatted();
 
         //- Write out variables for debugging
         virtual void writeVariables();
-
-        //- Update for changes of mesh
-        virtual void updateMesh(const mapPolyMesh&)
-        {}
-
-        //- Update for changes of mesh
-        virtual void movePoints(const polyMesh&)
-        {}
 };
