--- src/postProcessing/functionObjects/utilities/spinnerLidar/spinnerLidarFunctionObject.H
@@ -1,54 +1,0 @@
-/*---------------------------------------------------------------------------*\
-  =========                 |
-  \\      /  F ield         | OpenFOAM: The Open Source CFD Toolbox
-   \\    /   O peration     |
-    \\  /    A nd           | Copyright (C) 2012 OpenFOAM Foundation
-     \\/     M anipulation  |
--------------------------------------------------------------------------------
-License
-    This file is part of OpenFOAM.
-
-    OpenFOAM is free software: you can redistribute it and/or modify it
-    under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    OpenFOAM is distributed in the hope that it will be useful, but WITHOUT
-    ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
-    FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
-    for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with OpenFOAM.  If not, see <http://www.gnu.org/licenses/>.
-
-Typedef
-    Foam::spinnerLidarFunctionObject
-
-Description
-    FunctionObject wrapper around spinnerLidar to allow it to be
-    created via the functions entry within controlDict.
-
-SourceFiles
-    spinnerLidarFunctionObject.C
-
-\*---------------------------------------------------------------------------*/
-
-#ifndef spinnerLidarFunctionObject_H
-#define spinnerLidarFunctionObject_H
-
-#include "spinnerLidar.H"
-#include "OutputFilterFunctionObject.H"
-
-// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
-
-namespace Foam
-{
-    typedef OutputFilterFunctionObject<spinnerLidar>
-        spinnerLidarFunctionObject;
-}
-
-// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
-
-#endif
-
-// ************************************************************************* //
